+++ esp-hal/src/interrupt/mod.rs
@@ -73,8 +73,6 @@ We reserve a number of CPU interrupts, which cannot be used; see
 //! }
 //! ```
 
-use core::num::NonZeroUsize;
-
 #[cfg(riscv)]
 pub use self::riscv::*;
 #[cfg(xtensa)]
@@ -144,54 +142,32 @@ pub trait InterruptConfigurable: crate::private::Sealed {
 #[derive(Copy, Clone, Debug)]
 #[cfg_attr(feature = "defmt", derive(defmt::Format))]
 pub struct IsrCallback {
-    f: NonZeroUsize,
+    f: extern "C" fn(),
 }
 
 impl IsrCallback {
     /// Construct a new callback from the callback function.
     pub fn new(f: extern "C" fn()) -> Self {
         // a valid fn pointer is non zero
-        Self {
-            f: unwrap!(NonZeroUsize::new(f as usize)),
-        }
-    }
-
-    /// Construct a new callback from the callback function and the nested flag.
-    pub(crate) fn new_with_nested(f: extern "C" fn(), nested: bool) -> Self {
-        // a valid fn pointer is non zero
-        let f = unwrap!(NonZeroUsize::new(f as usize | !nested as usize));
         Self { f }
     }
 
-    /// Construct a new callback from a raw value.
-    ///
-    /// # Panics
-    ///
-    /// Passing zero is invalid and results in a panic.
-    pub fn from_raw(f: usize) -> Self {
-        Self {
-            f: unwrap!(NonZeroUsize::new(f)),
-        }
-    }
-
-    /// Returns the raw value of the callback.
-    ///
-    /// Don't just cast this to function and call it - it might be misaligned.
-    pub fn raw_value(self) -> usize {
-        self.f.into()
+    /// Returns the address of the callback.
+    pub fn address(self) -> usize {
+        self.f as usize
     }
 
     /// The callback function.
     ///
     /// This is aligned and can be called.
-    pub fn aligned_ptr(self) -> extern "C" fn() {
-        unsafe { core::mem::transmute::<usize, extern "C" fn()>(Into::<usize>::into(self.f) & !1) }
+    pub fn callback(self) -> extern "C" fn() {
+        self.f
     }
 }
 
 impl PartialEq for IsrCallback {
     fn eq(&self, other: &Self) -> bool {
-        core::ptr::fn_addr_eq(self.aligned_ptr(), other.aligned_ptr())
+        core::ptr::fn_addr_eq(self.callback(), other.callback())
     }
 }
 
@@ -205,48 +181,19 @@ impl PartialEq for IsrCallback {
 pub struct InterruptHandler {
     f: extern "C" fn(),
     prio: Priority,
-    #[cfg(riscv)]
-    nested: bool,
 }
 
 impl InterruptHandler {
     /// Creates a new [InterruptHandler] which will call the given function at
     /// the given priority.
     pub const fn new(f: extern "C" fn(), prio: Priority) -> Self {
-        Self {
-            f,
-            prio,
-            #[cfg(riscv)]
-            nested: true,
-        }
-    }
-
-    /// Creates a new [InterruptHandler] which will call the given function at
-    /// the given priority with disabled interrupt nesting.
-    ///
-    /// Usually higher priority interrupts get served while handling an interrupt.
-    /// Using this the interrupt handler won't get preempted by higher priority interrupts.
-    #[cfg(riscv)]
-    pub fn new_not_nested(f: extern "C" fn(), prio: Priority) -> Self {
-        Self {
-            f,
-            prio,
-            nested: false,
-        }
+        Self { f, prio }
     }
 
     /// The Isr callback.
     #[inline]
     pub fn handler(&self) -> IsrCallback {
-        cfg_if::cfg_if! {
-            if #[cfg(riscv)] {
-                let nested = self.nested;
-            } else {
-                let nested = true;
-            }
-        }
-
-        IsrCallback::new_with_nested(self.f, nested)
+        IsrCallback::new(self.f)
     }
 
     /// Priority to be used when registering the interrupt